style: move kotlin code to appropriate subdirectory
