+++ kotlin/src/main/kotlin/de/frohnmeyerwds/mima/Main.kt
@@ -1,4 +1,4 @@
-package de.frohnmeyer_wds
+package de.frohnmeyerwds.mima
 
 import kotlin.io.path.Path
