+++ kotlin/src/main/kotlin/de/frohnmeyerwds/mima/U24Array.kt
@@ -1,4 +1,4 @@
-package de.frohnmeyer_wds
+package de.frohnmeyerwds.mima
 
 class U24Array constructor(val value: IntArray) {
     constructor(size: Int) : this(IntArray(size))
